requirements.txt
django==6.0.4
psycopg[binary,pool]==3.3.3
celery==5.4.0
django-celery-results==2.5.1
requests==2.32.3
sentry-sdk==2.58.0
uWSGI==2.0.31
pre-commit==4.5.1
freezegun==1.5.5
django-debug-toolbar==6.3.0
pytest-django==4.12.0
markdown==3.10.2
djangorestframework==3.17.1
django-filter==25.2
Pillow==12.2.0
django-storages[s3]==1.14.6
djangorestframework-simplejwt==5.5.1
django-cors-headers==4.9.0
